fixed an integer division error (should have been a float division)
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@567813 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/commons/math/util/ResizableDoubleArray.java
@@ -667,7 +667,7 @@ public class ResizableDoubleArray implements DoubleArray, Serializable {
      */
     private synchronized boolean shouldContract() {
         if (expansionMode == MULTIPLICATIVE_MODE) { 
-            return (internalArray.length / numElements) > contractionCriteria;
+            return (internalArray.length / ((float) numElements)) > contractionCriteria;
         } else {
             return (internalArray.length - numElements) > contractionCriteria;
         }
